ci!: Remove k8s-image kind and make push-image production-ready
