+++ taskcluster/docker/skopeo/push_image.sh
@@ -2,7 +2,8 @@
 set -e
 
 test $APP
-test $APP_VERSION
+test $DOCKER_TAG
+test $DOCKER_ARCHIVE_TAG
 test $DOCKER_REPO
 test $MOZ_FETCHES_DIR
 test $TASKCLUSTER_ROOT_URL
@@ -25,7 +26,6 @@ skopeo copy docker-archive:image.tar oci:${APP}:final
 cat > version.json <<EOF
 {
     "commit": "${VCS_HEAD_REV}",
-    "version": "${APP_VERSION}",
     "source": "${VCS_HEAD_REPOSITORY}",
     "build": "${TASKCLUSTER_ROOT_URL}/tasks/${TASK_ID}"
 }
@@ -34,10 +34,9 @@ EOF
 umoci insert --image ${APP}:final version.json /app/version.json
 
 echo "=== Pushing to docker hub ==="
-DOCKER_TAG="v${APP_VERSION}"
-# TODO: Enable pushing
-# skopeo copy oci:${APP}:final docker://$DOCKER_REPO:$DOCKER_TAG
-# skopeo inspect docker://$DOCKER_REPO:$DOCKER_TAG
+skopeo copy oci:${APP}:final docker://$DOCKER_REPO:$DOCKER_TAG
+skopeo copy oci:${APP}:final docker://$DOCKER_REPO:$DOCKER_ARCHIVE_TAG
+skopeo inspect docker://$DOCKER_REPO:$DOCKER_TAG
 
 echo "=== Clean up ==="
 rm -rf $HOME/.docker
