--- taskcluster/kinds/k8s-image/kind.yml
@@ -1,57 +1,0 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
----
-loader: taskgraph.loader.transform:loader
-
-kind-dependencies:
-    - tox
-
-transforms:
-    - scriptworker_taskgraph.transforms.python_version:transforms
-    - scriptworker_taskgraph.transforms.k8s_image:transforms
-    - scriptworker_taskgraph.transforms.cached:transforms
-    - taskgraph.transforms.cached_tasks:transforms
-    - taskgraph.transforms.task:transforms
-
-task-defaults:
-    description: "{name} k8s image builder py{python_version}"
-    name: "{name}-python{python_version}"
-    run-on-tasks-for: ["action", "github-pull-request", "github-push"]
-    attributes:
-        code-review: true
-    worker-type: images
-    worker:
-        docker-image: mozillareleases/k8s-image:3a46d79c28b31bbfe7d35142b3b1e621d7907ded@sha256:5db40915dd4d34f8fde4aceca3b398e0c954513b57d252bd7d5d2833ae41d2cc
-        max-run-time: 3600
-        docker-in-docker: true
-        taskcluster-proxy: true
-        chain-of-trust: true
-        artifacts:
-            - type: directory
-              name: public/
-              path: /builds/worker/artifacts/
-        command:
-            - sh
-            - -lxce
-            - >-
-              sh /usr/local/bin/build_and_push.sh
-    python-versions:
-        - 39
-    deploy-secret-url: http://taskcluster/secrets/v1/secret/project/releng/scriptworker-scripts/deploy
-    docker-repo: mozilla/releng-{name}
-    # TODO copy image to artifacts
-
-tasks:
-    addonscript: {}
-    balrogscript: {}
-    beetmoverscript: {}
-    bitrisescript: {}
-    bouncerscript: {}
-    githubscript: {}
-    pushapkscript: {}
-    pushflatpakscript: {}
-    pushmsixscript: {}
-    shipitscript: {}
-    signingscript: {}
-    treescript: {}
