+++ taskcluster/kinds/pr/kind.yml
@@ -6,7 +6,7 @@ loader: taskgraph.loader.transform:loader
 
 kind-dependencies:
     - tox
-    - k8s-image
+    - docker-image
 
 transforms:
     - taskgraph.transforms.code_review:transforms
