+++ taskcluster/scriptworker_taskgraph/target_tasks.py
@@ -18,4 +18,4 @@ def target_tasks_default(full_task_graph, parameters, graph_config):
 @register_target_task("docker-hub-push")
 def target_tasks_docker_hub_push(full_task_graph, parameters, graph_config):
     """Filter by kind and `script-name`."""
-    return [l for l, t in full_task_graph.tasks.items() if t.kind == "k8s-image" and filter_for_script(t, parameters)]  # noqa: E741
+    return [l for l, t in full_task_graph.tasks.items() if t.kind == "push-image" and filter_for_script(t, parameters)]  # noqa: E741
